refactor(server): general cleanup
