+++ src/config/h5p-config.ts
@@ -1,14 +1,17 @@
 export default {
-    uuid: '8de62c47-f335-42f6-909d-2d8f4b7fb7f5',
-    sendUsageStatistics: false,
     ajaxUrl: '/ajax',
     baseUrl: '/api/h5p/v1',
+    contentFilesUrl: '/content',
+    coreUrl: '/core',
     downloadUrl: '/download',
     editorLibraryUrl: '/editor',
+    exportMaxContentPathLength: 255,
     librariesUrl: '/libraries',
-    coreUrl: '/core',
-    contentFilesUrl: '/content',
+    maxFileSize: 2000 * 1024 * 1024,
+    maxTotalSize: 2000 * 1024 * 1024,
     platformName: 'Lumi H5P Editor',
     platformVersion: '0.1',
-    temporaryFilesUrl: '/tmp'
+    sendUsageStatistics: false,
+    temporaryFilesUrl: '/tmp',
+    uuid: '8de62c47-f335-42f6-909d-2d8f4b7fb7f5'
 };
