+++ src/server/boot/setup.ts
@@ -1,19 +1,15 @@
 import mkdirp from 'mkdirp';
+import * as Sentry from '@sentry/node';
 
 import appConfig from '../config/app-config';
 
-import * as Sentry from '@sentry/node';
-
-export default function setup(): Promise<void> {
-    return new Promise((resolve, reject) => {
-        try {
-            mkdirp.sync(appConfig.workingCachePath);
-            mkdirp.sync(appConfig.librariesPath);
-            mkdirp.sync(appConfig.temporaryStoragePath);
-            resolve();
-        } catch (error) {
-            Sentry.captureException(error);
-            reject(error);
-        }
-    });
+export default async function setup(): Promise<void> {
+    try {
+        mkdirp.sync(appConfig.workingCachePath);
+        mkdirp.sync(appConfig.librariesPath);
+        mkdirp.sync(appConfig.temporaryStoragePath);
+    } catch (error) {
+        Sentry.captureException(error);
+        throw error;
+    }
 }
